refactor: posts list refactoring
