+++ src/entities/posts-list/api/postListApi.graphql
