+++ src/entities/posts-list/ui/PostsList.module.scss
@@ -1,8 +1,6 @@
 .posts {
   display: flex;
   flex-wrap: wrap;
-  gap: 12px;
-  justify-content: space-around;
-
+  gap: 1rem;
   margin: 1.5rem 0 2.25rem;
 }
