+++ src/entities/posts-list/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './PostsList'
